Remove boilerplate Button component files and stub out Chakra Button for styling
