+++ src/Typography.mdx
@@ -1,8 +1,6 @@
 import { Meta, Unstyled } from '@storybook/blocks';
-import { tokens } from "../theme/tokens"
-import { ThemeBlock } from "../../.storybook/ThemeBlock"
+import { ThemeBlock } from "../.storybook/ThemeBlock"
 import { Text, HStack, VStack } from "@chakra-ui/react";
-import { theme } from "../theme";
 
 <Meta title="Tokens/Typography" />
