+++ src/components/Button/Button.stories.ts
@@ -1,0 +1,25 @@
+import type { Meta, StoryObj } from "@storybook/react";
+import { getThemingArgTypes } from "@chakra-ui/storybook-addon";
+import { Button } from "./Button";
+import { theme } from "../../theme";
+
+const meta = {
+  title: "Components/Button",
+  component: Button,
+  tags: ["autodocs"],
+  argTypes: {
+    ...getThemingArgTypes(theme, "Button"),
+    children: { type: "string" },
+  },
+} satisfies Meta<typeof Button>;
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const Primary: Story = {
+  args: {
+    variant: "primary",
+    size: "md",
+    children: "Click here",
+  },
+};
