+++ src/components/Button/Button.tsx
@@ -1,0 +1,2 @@
+export { Button } from "@chakra-ui/react";
+export type { ButtonProps } from "@chakra-ui/react";
