+++ src/components/Button/index.ts
@@ -1,0 +1,2 @@
+export { Button } from "./Button";
+export type { ButtonProps } from "./Button";
